backend/pyproject.toml
[project]
name = "backend-starter"
version = "1.0.0"
description = "Backend for stater template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.2.1",
    "pytest-asyncio==0.24.0",
    "litestar>=2.15.2",
    "pydantic-settings>=2.8.1",
    "pydantic>=2.11.3",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
    "advanced-alchemy==1.4.4",
    "bcrypt>=4.2.1",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "asyncpg>=0.30.0",
    "pyotp>=2.9.0",
    "time-machine>=2.19.0",
]

[tool.ruff]
builtins = ["_"]
line-length = 120

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.isort]
detect-same-package = true
force-sort-within-sections = true
